dmipy_test_project/dmipy_py36_requirements.txt
absl-py==0.9.0
appdirs==1.4.3
astor==0.8.1
boto==2.49.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
cvxpy==1.0.25
cycler==0.10.0
dill==0.3.1.1
dipy==1.1.1
distlib==0.3.0
dmipy==0.1.dev0
ecos==2.0.7.post1
filelock==3.0.12
fury==0.4.0
future==0.18.2
gast==0.2.2
google-auth==1.10.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
graphviz==0.13.2
grpcio==1.26.0
h5py==2.10.0
idna==2.8
importlib-metadata==1.5.0
importlib-resources==1.0.2
joblib==0.14.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
llvmlite==0.31.0
Markdown==3.1.1
matplotlib==3.2.0rc1
multiprocess==0.70.9
nibabel==3.0.0
numba==0.47.0
numpy==1.18.1
oauthlib==3.1.0
opt-einsum==3.1.0
osqp==0.6.1
pandas==1.0.0rc0
pathlib==1.0.1
pathos==0.2.5
pox==0.2.7
ppft==1.6.6.1
protobuf==3.11.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
scs==2.1.1.post2
seaborn==0.9.0
six==1.14.0
tensorboard==1.15.0
tensorflow-estimator==1.15.1
tensorflow-gpu==1.15.0
termcolor==1.1.0
urllib3==1.25.8
virtualenv==20.0.3
vtk==8.1.2
Werkzeug==0.16.0
wrapt==1.11.2
zipp==2.2.0
